+++ src/main/java/lumaceon/mods/clockworkphase2/api/time/ItemTimeContainer.java
@@ -7,9 +7,9 @@ import net.minecraft.item.ItemStack;
 
 public class ItemTimeContainer extends Item implements ITimeContainerItem
 {
-    public long capacity;
+    public int capacity;
 
-    public ItemTimeContainer(int maxStack, int maxDamage, long capacity)
+    public ItemTimeContainer(int maxStack, int maxDamage, int capacity)
     {
         this.setMaxStackSize(maxStack);
         this.setMaxDamage(maxDamage);
@@ -17,36 +17,36 @@ public class ItemTimeContainer extends Item implements ITimeContainerItem
     }
 
     @Override
-    public long receiveTime(ItemStack timeItem, long maxReceive, boolean simulate)
+    public int receiveTime(ItemStack timeItem, int maxReceive, boolean simulate)
     {
-        long currentTime = NBTHelper.LONG.get(timeItem, NBTTags.TIME);
-        long timeReceived = Math.min(getEmptySpace(timeItem), maxReceive);
+        int currentTime = NBTHelper.INT.get(timeItem, NBTTags.TIME);
+        int timeReceived = Math.min(getEmptySpace(timeItem), maxReceive);
         if(!simulate)
-            NBTHelper.LONG.set(timeItem, NBTTags.TIME, currentTime + timeReceived);
+            NBTHelper.INT.set(timeItem, NBTTags.TIME, currentTime + timeReceived);
         return timeReceived;
     }
 
     @Override
-    public long extractTime(ItemStack timeItem, long maxExtract, boolean simulate)
+    public int extractTime(ItemStack timeItem, int maxExtract, boolean simulate)
     {
-        long currentTime = NBTHelper.LONG.get(timeItem, NBTTags.TIME);
-        long timeExtracted = Math.min(currentTime, maxExtract);
+        int currentTime = NBTHelper.INT.get(timeItem, NBTTags.TIME);
+        int timeExtracted = Math.min(currentTime, maxExtract);
         if(!simulate)
-            NBTHelper.LONG.set(timeItem, NBTTags.TIME, currentTime - timeExtracted);
+            NBTHelper.INT.set(timeItem, NBTTags.TIME, currentTime - timeExtracted);
         return timeExtracted;
     }
 
     @Override
-    public long getMaxCapacity(ItemStack timeItem) {
+    public int getMaxCapacity(ItemStack timeItem) {
         return capacity;
     }
 
     @Override
-    public long getTimeStored(ItemStack timeItem) {
-        return NBTHelper.LONG.get(timeItem, NBTTags.TIME);
+    public int getTimeStored(ItemStack timeItem) {
+        return NBTHelper.INT.get(timeItem, NBTTags.TIME);
     }
 
-    public long getEmptySpace(ItemStack timeItem) {
+    public int getEmptySpace(ItemStack timeItem) {
         return getMaxCapacity(timeItem) - getTimeStored(timeItem);
     }
 }
